refactor: decompose page-tree.tsx into focused sub-components (#690)
