+++ .agents/architecture.md
@@ -400,7 +400,10 @@ src/
 │   │   ├── create-workspace-dialog.tsx # Dialog for creating a new workspace
 │   │   ├── page-search.tsx      # Full-text search input + results dropdown (debounced, 300ms)
 │   │   ├── favorites-section.tsx # Per-user favorites list + useFavorite hook for toggle
-│   │   ├── page-tree.tsx        # Hierarchical page tree with CRUD, drag-and-drop, nest/unnest, favorites toggle (uses lib/page-tree.ts)
+│   │   ├── page-tree.tsx        # Orchestrator: data fetching, state, delete dialog (uses extracted sub-modules)
+│   │   ├── page-tree-item.tsx  # Single tree node rendering + context menu
+│   │   ├── page-tree-drag-layer.ts # usePageTreeDrag hook: drag-and-drop state + handlers
+│   │   ├── use-page-tree-actions.ts # usePageTreeActions hook: CRUD operations (create, delete, duplicate, move, nest, favorites)
 │   │   ├── trash-section.tsx    # Trash bin: lists soft-deleted pages, restore, permanent delete, empty trash
 │   │   └── user-menu.tsx        # User dropdown with settings link + sign-out
 │   ├── editor/                  # Lexical block editor
